+++ boost_manager/tests/common/mod.rs
@@ -2,13 +2,7 @@ use file_store::file_sink::{FileSinkClient, Message as SinkMessage};
 use helium_proto::BoostedHexInfoV1 as BoostedHexInfoProto;
 use helium_proto::BoostedHexUpdateV1 as BoostedHexUpdateProto;
 use helium_proto::Message;
-use mobile_config::boosted_hex_info::BoostedHexInfo;
 use tokio::{sync::mpsc::error::TryRecvError, time::timeout};
-
-#[derive(Debug, Clone)]
-pub struct MockHexBoostingClient {
-    pub boosted_hexes: Vec<BoostedHexInfo>,
-}
 
 pub struct MockFileSinkReceiver {
     pub receiver: tokio::sync::mpsc::Receiver<SinkMessage>,
